+++ src/settings.ts
@@ -24,6 +24,7 @@ export class VisualSettings extends DataViewObjectsParser implements IVisualSett
 
 export class ViewSettings {
     public colorMode: string = "light";
+    public enableCopyMenu: boolean = false;
 }
 
 export class MermaidSettings {